requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
black==24.4.2
blessed==1.20.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.7.0
cryptography==44.0.2
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-picklefield==3.3
django-q2==1.7.6
djangorestframework==3.15.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
exponent_server_sdk==2.1.0
filelock==3.18.0
flake8==7.0.0
fsspec==2025.3.2
google-auth==2.38.0
google-genai==1.10.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
jinxed==1.3.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
pdfminer.six==20250327
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.2.0
pytest==8.2.1
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
# sentence-transformers==4.0.2 # Now installed via Dockerfile
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
# torch==2.6.0 # Replaced with CPU-only version below
tqdm==4.67.1
# transformers==4.51.2 # Now installed via Dockerfile
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.6.0
pypdf
python-docx
pgvector
langchain
# sentence-transformers # Now installed via Dockerfile
# torch --index-url https://download.pytorch.org/whl/cpu # Now installed via Dockerfile